Revert "Merge remote-tracking branch 'upstream/main'"
This reverts commit fb03419dfc68b872de537fd4a196267e5e024a16, reversing
changes made to 35e7efd00c1dc4e29b1f120db2e1a3f46ac06525.

+++ source/isaaclab/isaaclab/sensors/camera/tiled_camera.py
@@ -248,9 +248,7 @@ class TiledCamera(Camera):
 
             # convert data buffer to warp array
             if isinstance(tiled_data_buffer, np.ndarray):
-                # Let warp infer the dtype from numpy array instead of hardcoding uint8
-                # Different annotators return different dtypes: RGB(uint8), depth(float32), segmentation(uint32)
-                tiled_data_buffer = wp.array(tiled_data_buffer, device=self.device)
+                tiled_data_buffer = wp.array(tiled_data_buffer, device=self.device, dtype=wp.uint8)
             else:
                 tiled_data_buffer = tiled_data_buffer.to(device=self.device)
